+++ src/explore-education-statistics-admin/src/services/common/service.ts
@@ -1,6 +1,6 @@
 import {IdLabelPair} from "@admin/services/common/types";
 import {createClient} from '@admin/services/util/service';
-import mocks from './mock/axios-mock';
+import mocks from './mock/mock-service';
 
 const apiClient = createClient({
   mockBehaviourRegistrar: mocks,
