+++ src/explore-education-statistics-admin/src/services/edit-release/setup/service.ts
@@ -1,6 +1,6 @@
 import {ReleaseSetupDetails, ReleaseSetupDetailsUpdateRequest,} from '@admin/services/edit-release/setup/types';
 import {createClient} from '@admin/services/util/service';
-import mocks from './mock/axios-mock';
+import mocks from './mock/mock-service';
 
 const apiClient = createClient({
   mockBehaviourRegistrar: mocks,
